+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 - Fastify backend
 - Winston logger
 - Vite / Vue 3 frontend
+- Serve frontend static files in production